amazon-s3 Buckect Update

+++ app/Http/Controllers/HomeBladeController.php
@@ -178,8 +178,11 @@ class HomeBladeController extends Controller
 
              //Delete the Old IMAGE AVATAR from  avatar Folder (Save Space)
              $imagePath = explode('/', $user_id->avatar);
-             $avatarName = $imagePath[5];
-             Storage::disk('s3')->delete('Uploads/avatars/'.$avatarName);
+             if ($imagePath[5] != NULL) {
+               $avatarName = $imagePath[5];
+               Storage::disk('s3')->delete('Uploads/avatars/'.$avatarName);
+             }
+
 
              //store another avatar
              $user_image = request()->file('avatar');
